WIP to fix race condition which can cause timezone warnings to show in places they shouldn't.
