+++ src/project/utils/decorators.py
@@ -1,0 +1,7 @@
+def with_attrs(**attrs):
+    def with_attrs(f):
+        for k,v in attrs.items():
+            setattr(f, k, v)
+        return f
+
+    return with_attrs
